requirements.txt
pytest==7.2
allure-pytest==2.12
selene==2.0.0b16
python-dotenv=1.0.0